Fixed issue where Digits and Twitter sessions are lost
If the name of TwitterCore or Digits class is modified by dexguard
or proguard, then it's possible that the shared preference name will
differ between build versions. As a result, when the PreferenceStoreImpl
tries to load the shared preference file, it cannot be found.

This commit will migrate existing shared preference file to a new file
created using a string that will not be obfuscated by Proguard or
Dexguard.

Jira-Id: TFW-5458
Change-Id: I71a4aba0f78315abff9cacc4c12afad9126784de