Fix memory leak when loading Tweet images
Problem:
Picasso keeps a strong reference to the success/error callback which inturn
holds a strong reference to the ImageView. This can lead to temporary memory
leak of the Activity if all requests have not completed before activity is
destroyed.

Solution:
Hold a weak reference to ImageView in the Callback.

Fixes: https://github.com/twitter/twitter-kit-android/issues/55
Change-Id: I3f99d1d4e1194d150edbe77e39c472df51e2cec1