Recording and retrieving invites
Problem:
--------
Invites sent by a user are not recorded for retrieval later on.

Solution:
--------
In this change, we wire up the enpoints to read and write to the invites
store of a Digits user.

Design decisions:
We make the invites status of a contact available via the cursor pointing
to the contact. This leaves room to expose this functionality to an app
in the future if we decide to.

When a contact is selected to be invited, we provide instant UI feedback
via the button being pressed. We also wire through the newly created
invite into the Contacts cursor without requiring a fetch from the
service.

Although the invites read enpoint exposes several pieces of metadata, we
currently only inflate those that are necessary.

Appendix
--------
In addition to the above change, we also enable logged out users use the
invite flow.

Change-Id: Id609cdf589bc7a4a49dc5d653ad7ffcbe2ea638e